Added new instructions: 	- ClrCb 	- ICountInt64_X 	- RICountInt64_X
For what these instructions do, please refer to the documentations on Instructions.

Changed the instruction counter for instruction counter interrupt to 64 bits from 32 bits.

Changed the return type for interrupt handler function to 1 from 0 when interrupt is disabled.
This means that a disabled interrupt is always considered as successfully handled.

Added interpreter instructions documentation in documentations folder.